requirements.txt
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.8
beanie==1.29.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.17.0
flake8==7.1.2
graphql-core==3.2.6
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.8
idna==3.10
isort==6.0.1
Jinja2==3.1.5
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
pycodestyle==2.12.1
pydantic==2.10.6
pydantic_core==2.27.2
pyflakes==3.2.0
Pygments==2.19.1
pylint==3.3.4
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.0
strawberry-graphql==0.262.0
toml==0.10.2
tomlkit==0.13.2
typer==0.15.2
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.29.2
watchfiles==1.0.4
websockets==15.0
